backend/requirements.txt
fastapi
uvicorn
redis==6.2.0
pyjwt
passlib[bcrypt]
python-dotenv
python-jose
pydantic
pydantic-core
debugpy
ffmpeg
yt_dlp
aiohttp
aiofiles